build: rename lucia files and fix import
