+++ lucia.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="lucia-auth" />
 declare namespace Lucia {
-	type Auth = import("./lib/lucia.js").Auth;
-	type UserAttributes = {
-		username: string;
-	};
+  type Auth = import("./lib-server/lucia.ts").Auth;
+  type UserAttributes = {
+    username: string;
+  };
 }
