+++ pages/sign-in.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { auth } from "../auth/lucia";
+import { auth } from "../lib-server/lucia";
 import { useRouter } from "next/router";
 
 import type { GetServerSidePropsContext, GetServerSidePropsResult } from "next";
